+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Updates/Fixes
 
 - Fix UTXO send price fetch [256](https://github.com/asgardex/asgardex-desktop/pull/257)
+- Fix Maya lp Bugs & BSC.USDT [#255](https://github.com/asgardex/asgardex-desktop/pull/258)
 
 # 1.21.7 (2024-05-23)