+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - Fix UTXO send price fetch [256](https://github.com/asgardex/asgardex-desktop/pull/257)
 - Fix Maya lp Bugs & BSC.USDT [#255](https://github.com/asgardex/asgardex-desktop/pull/258)
-- Update Thornode Leedger to use ClientLedger from xchainjs [#259](https://github.com/asgardex/asgardex-desktop/pull/260)
+- Update Thornode Ledger to use ClientLedger from xchainjs [#259](https://github.com/asgardex/asgardex-desktop/pull/260)
+- Disable Unbonding if node is active [#233] (https://github.com/asgardex/asgardex-desktop/pull/263)
 
 # 1.21.7 (2024-05-23)